[MOD] almost there ! trying to 'normalize' everything now, so everything is as clean as possible

+++ src/expand/matching.c
@@ -6,19 +6,20 @@
 /*   By: cviel <cviel@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/16 15:53:48 by cviel             #+#    #+#             */
-/*   Updated: 2025/08/04 17:41:41 by cviel            ###   ########.fr       */
+/*   Updated: 2025/08/05 18:27:26 by cviel            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <unistd.h>
+#include <stdlib.h>
 #include "minishell.h"
 #include "expand.h"
 #include "ft_vectors.h"
 #include "ft_string.h"
 
-static int	build_segment(char **seg, t_vector pattern, int *ind);
+static int	build_segment(char **seg, t_vector pattern, size_t *ind);
 static int	matching_start(t_vector pattern, char **p_elem,
-				int *match, int *ind);
+				int *match, size_t *ind);
 static void	matching_segment(char **p_elem, char **seg, int *p_match);
 static void	matching_end(char *elem, char *seg, int *p_match);
 
@@ -26,19 +27,19 @@ int	matching(t_vector pattern, char *elem, int *p_match)
 {
 	int		ret;
 	char	*seg;
-	int		i;
+	size_t	i;
 
 	i = 0;
 	ret = matching_start(pattern, &elem, p_match, &i);
 	if (ret != 0 || *p_match == FALSE)
 		return (ret);
 	seg = NULL;
-	while ((size_t)i < pattern.size)
+	while (i < pattern.size)
 	{
 		ret = build_segment(&seg, pattern, &i);
 		if (ret != 0 || seg == NULL)
 			return (ret);
-		if ((size_t)i < pattern.size)
+		if (i < pattern.size)
 		{
 			matching_segment(&elem, &seg, p_match);
 			if (*p_match == FALSE)
@@ -51,14 +52,14 @@ int	matching(t_vector pattern, char *elem, int *p_match)
 	return (0);
 }
 
-static int	build_segment(char **seg, t_vector pattern, int *ind)
+static int	build_segment(char **seg, t_vector pattern, size_t *ind)
 {
 	char	*join;
 
-	while ((size_t)(*ind) < pattern.size
+	while (*ind < pattern.size
 		&& ((t_pat *)pattern.data)[*ind].wild == TRUE)
 		(*ind)++;
-	while ((size_t)(*ind) < pattern.size
+	while (*ind < pattern.size
 		&& ((t_pat *)pattern.data)[*ind].wild == FALSE)
 	{
 		join = malloc(sizeof(char) * 2);
@@ -76,7 +77,7 @@ static int	build_segment(char **seg, t_vector pattern, int *ind)
 }
 
 static int	matching_start(t_vector pattern,
-				char **p_elem, int *match, int *ind)
+				char **p_elem, int *match, size_t *ind)
 {
 	int		ret;
 	char	*seg;
